1. SQL script minor change. 2. Agreement form now validates input values. 3. Added <!DOCTYPE html> (minor) 4. AgreementRequestBody int to double

+++ src/main/java/lv/velexauto/velex/HelperClasses/AgreementRequestBody.java
@@ -16,7 +16,7 @@ public class AgreementRequestBody {
     String onBehalfOf;
     String fileLinkAgreement;
     String fileLinkInvoice;
-    int price;
+    double price;
     double valueAddedTax;
     int paymentTerm;
     String driver;
@@ -64,7 +64,7 @@ public class AgreementRequestBody {
         return fileLinkInvoice;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
